requirements.txt
selenium
webdriver_manager
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib